manageme/pyproject.toml
[tool.poetry]
name = "shue-store"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
psycopg2 = "^2.9.9"
django-bootstrap-modal-forms = "^3.0.4"
django-widget-tweaks = "^1.5.0"
django-allauth = "^0.62.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
